Scrape amendment "house_number" so it corresponds with House roll call vote XML
As disovered in issue #68, the A___ numbering found on THOMAS (and entered by
House floor staff) does not correspond to the amendment-num field in House
roll call vote XML --- mostly when there is a House Rules report that has
an amendment in the nature of a substitute, which shifts the numbering of
Member amendments made in order by the rule.

The number used in vote XML is more or less scrapable from the purpose
(and if not purpose then sometimes description) field of the amendment. It
looks like "An amendment numbered XXX printed in (a report|the CR)."

This commit replaces the "house_number" field in our output (and the "sequence"
attribute in GovTrack-style XML output) with the scraped number as I've
just described so that the field now corresponds with House vote output.

The house_number field previously stored the A___ number, which is now
stored in a new field named "offered_order," since the A___ number is
consecutive in the order in which an amendment was offered.