When choosing {short,official,popular}_title, only choose from titles for the whole bill.
On THOMAS and Congress.gov, the main title displayed for a bill is never a title for
a portion of the bill, and if there is no title for the whole bill in the last 'as'
group, they display a title (for the whole bill) from the previous 'as' group

This affects four bills in the 113th Congress (so far). Three bills now on longer
have a short_title (it's now null), and one's short title changed (HR 1911).