improve assigning committee IDs in action lines when the committee's chamber is ambiguous, fixes #108
Don't assume that the committee is in the bill's originating chamber. If it's a committee action,
look at the chamber of the committee the action is taking place in. Failing that, use some specific
regular expressions to see if it is a House or Senate action. And failing that, if the bill is
in an early stage when we are pretty sure actions are in the originating chamber use the originating
chamber.

This makes a number of corrections, but also some action lines lose their committee IDs. In some
cases (lots of references to the Budget Act) the original ID was incorrect. In other cases it's
ambiguous or hard to figure out.

also see #110