have ENACTED:SIGNED be triggered at 'Signed by President', add a new code ENACTED:TENDAYRULE
Per #106, ENACTED:SIGNED should be triggered when the President signs a bill, which is
when the bill seems to actually become law. The "Became Public Law" action is typically
dated the same but may not actually be posted until much later, when the Office of the
Federal Register assigns the public law number. This can be problematic when trying to
count laws.

The same problem might occur with vetoed bills. In principle they must also become law
when the second chamber finishes its override, or thereabouts, but our ENACTED:VETO_OVERRIDE
status is still triggered by Became Public Law. This is more rare so I'll punt this
for another time.

This left a gap for six bills (see the commit for the list) that had Became Public Law
actions but neither Signed by President nor veto actions. They appear to be instances
of the "ten Days (Sundays excepted)" provision in the Constitution. So here I'm also
adding a new status code called ENACTED:TENDAYRULE. Like the other conditions, surely
they become law on the actual 10th day regardless of whether OFR assigns them a number,
but that is harder to detect. There is a "Sent to Archivist of the United States unsigned"
action that we might want to use instead. But this is historical and incredibly rare
so it doesn't make much difference now.