add a --diff option for bills and votes to test if output has changed
When --diff is specified (for bills & votes), instead of writing output files to
disk, we run a diff over the existing file and the new content and display the
diff. This is handy for testing.

At the same time I'm removing my previous preserve_update_time flag which
I had been using for a similar purpose, but this new method is much easier.

reverts 5122ad6f966ba5899a0758ed92d81ca779314c7f