fdsys: package lastmod cache files weren't quite right
Each time a package was downloaded (--store ...) the lastmod time from
the sitemap was written to disk in lastmod.txt next to the package. But
since an invocation of the fdsys script can choose to store different
file types, we would lose track of the lastmods of some of the files if
fdsys was invoked again to --store different files.

This revises the lastmod cache strategy to store a .json file containing
the lastmod of each individual downloaded file (type), rather than for
the whole package.