"Placed on the ... Calendar" should not cause a bill to change state to REPORTED
Although this indicates that a bill is ready for floor action, in the Senate this may result from an opponent objecting to committee consideration, which seems to be a way to kill a bill:

* https://www.everycrsreport.com/reports/RS22309.html
* https://www.congress.gov/congressional-record/2022/5/10/senate-section/article/s2385-7?q=%7B%22search%22%3A%5B%22S.4164%22%2C%22S.4164%22%5D%7D&s=1&r=1

The "type" code (which no one uses) on actions that are actually for "ordered reported" are changed from "calendar" to "ordered-reported".

The "Committee ... Reported by" regex is revised to catch "original measure" and "to Senate".