Don't try to fetch votes from future sessions
In ea5ba20fba95cb6ded70e7c704f6c102857e3904, I revised the votes task to accept just a congress number and it would come up with the session numbers. But this led to unneecssary warnings about not being able to get votes for a session that hasn't yet started.